Remove the interface for DefaultStatsCollector and DefaultClientMetricReport, rename them to remove "Default" suffix (#2097)
* Remove StatsCollector interface

* Rename DefaultStatsCollector to StatsCollector

* Remove ClientMetricReport interface

* Rename DefaultClientMetricReport to ClientMetricReport

* Remove clientMetricReport parameter from StatsCollector.start() API

* Replace getRTCPeerConnectionStats with getRTCStatsReport in meeting demo
