+++ docs/classes/allhighestvideobandwidthpolicy.html
@@ -247,7 +247,7 @@
 					<a name="updatemetrics" class="tsd-anchor"></a>
 					<h3>update<wbr>Metrics</h3>
 					<ul class="tsd-signatures tsd-kind-method tsd-parent-kind-class">
-						<li class="tsd-signature tsd-kind-icon">update<wbr>Metrics<span class="tsd-signature-symbol">(</span>_clientMetricReport<span class="tsd-signature-symbol">: </span><a href="../interfaces/clientmetricreport.html" class="tsd-signature-type">ClientMetricReport</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">void</span></li>
+						<li class="tsd-signature tsd-kind-icon">update<wbr>Metrics<span class="tsd-signature-symbol">(</span>_clientMetricReport<span class="tsd-signature-symbol">: </span><a href="clientmetricreport.html" class="tsd-signature-type">ClientMetricReport</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">void</span></li>
 					</ul>
 					<ul class="tsd-descriptions">
 						<li class="tsd-description">
@@ -260,7 +260,7 @@
 							<h4 class="tsd-parameters-title">Parameters</h4>
 							<ul class="tsd-parameters">
 								<li>
-									<h5>_clientMetricReport: <a href="../interfaces/clientmetricreport.html" class="tsd-signature-type">ClientMetricReport</a></h5>
+									<h5>_clientMetricReport: <a href="clientmetricreport.html" class="tsd-signature-type">ClientMetricReport</a></h5>
 								</li>
 							</ul>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">void</span></h4>
