+++ docs/classes/audiovideocontrollerstate.html
@@ -529,7 +529,7 @@
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-class">
 					<a name="statscollector" class="tsd-anchor"></a>
 					<h3>stats<wbr>Collector</h3>
-					<div class="tsd-signature tsd-kind-icon">stats<wbr>Collector<span class="tsd-signature-symbol">:</span> <a href="../interfaces/statscollector.html" class="tsd-signature-type">StatsCollector</a><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">null</span><span class="tsd-signature-symbol"> = null</span></div>
+					<div class="tsd-signature tsd-kind-icon">stats<wbr>Collector<span class="tsd-signature-symbol">:</span> <a href="statscollector.html" class="tsd-signature-type">StatsCollector</a><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">null</span><span class="tsd-signature-symbol"> = null</span></div>
 					<aside class="tsd-sources">
 						<ul>
 							<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/main/src/audiovideocontroller/AudioVideoControllerState.ts#L124">src/audiovideocontroller/AudioVideoControllerState.ts:124</a></li>
