+++ docs/classes/signalingandmetricsconnectionmonitor.html
@@ -112,7 +112,7 @@
 					<a name="constructor" class="tsd-anchor"></a>
 					<h3>constructor</h3>
 					<ul class="tsd-signatures tsd-kind-constructor tsd-parent-kind-class">
-						<li class="tsd-signature tsd-kind-icon">new <wbr>Signaling<wbr>And<wbr>Metrics<wbr>Connection<wbr>Monitor<span class="tsd-signature-symbol">(</span>audioVideoController<span class="tsd-signature-symbol">: </span><a href="../interfaces/audiovideocontroller.html" class="tsd-signature-type">AudioVideoController</a>, realtimeController<span class="tsd-signature-symbol">: </span><a href="../interfaces/realtimecontroller.html" class="tsd-signature-type">RealtimeController</a>, videoTileController<span class="tsd-signature-symbol">: </span><a href="../interfaces/videotilecontroller.html" class="tsd-signature-type">VideoTileController</a>, connectionHealthData<span class="tsd-signature-symbol">: </span><a href="connectionhealthdata.html" class="tsd-signature-type">ConnectionHealthData</a>, pingPong<span class="tsd-signature-symbol">: </span><a href="../interfaces/pingpong.html" class="tsd-signature-type">PingPong</a>, statsCollector<span class="tsd-signature-symbol">: </span><a href="../interfaces/statscollector.html" class="tsd-signature-type">StatsCollector</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><a href="signalingandmetricsconnectionmonitor.html" class="tsd-signature-type">SignalingAndMetricsConnectionMonitor</a></li>
+						<li class="tsd-signature tsd-kind-icon">new <wbr>Signaling<wbr>And<wbr>Metrics<wbr>Connection<wbr>Monitor<span class="tsd-signature-symbol">(</span>audioVideoController<span class="tsd-signature-symbol">: </span><a href="../interfaces/audiovideocontroller.html" class="tsd-signature-type">AudioVideoController</a>, realtimeController<span class="tsd-signature-symbol">: </span><a href="../interfaces/realtimecontroller.html" class="tsd-signature-type">RealtimeController</a>, videoTileController<span class="tsd-signature-symbol">: </span><a href="../interfaces/videotilecontroller.html" class="tsd-signature-type">VideoTileController</a>, connectionHealthData<span class="tsd-signature-symbol">: </span><a href="connectionhealthdata.html" class="tsd-signature-type">ConnectionHealthData</a>, pingPong<span class="tsd-signature-symbol">: </span><a href="../interfaces/pingpong.html" class="tsd-signature-type">PingPong</a>, statsCollector<span class="tsd-signature-symbol">: </span><a href="statscollector.html" class="tsd-signature-type">StatsCollector</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><a href="signalingandmetricsconnectionmonitor.html" class="tsd-signature-type">SignalingAndMetricsConnectionMonitor</a></li>
 					</ul>
 					<ul class="tsd-descriptions">
 						<li class="tsd-description">
@@ -139,7 +139,7 @@
 									<h5>pingPong: <a href="../interfaces/pingpong.html" class="tsd-signature-type">PingPong</a></h5>
 								</li>
 								<li>
-									<h5>statsCollector: <a href="../interfaces/statscollector.html" class="tsd-signature-type">StatsCollector</a></h5>
+									<h5>statsCollector: <a href="statscollector.html" class="tsd-signature-type">StatsCollector</a></h5>
 								</li>
 							</ul>
 							<h4 class="tsd-returns-title">Returns <a href="signalingandmetricsconnectionmonitor.html" class="tsd-signature-type">SignalingAndMetricsConnectionMonitor</a></h4>
@@ -200,7 +200,7 @@
 					<a name="metricsdidreceive" class="tsd-anchor"></a>
 					<h3>metrics<wbr>Did<wbr>Receive</h3>
 					<ul class="tsd-signatures tsd-kind-method tsd-parent-kind-class">
-						<li class="tsd-signature tsd-kind-icon">metrics<wbr>Did<wbr>Receive<span class="tsd-signature-symbol">(</span>clientMetricReport<span class="tsd-signature-symbol">: </span><a href="../interfaces/clientmetricreport.html" class="tsd-signature-type">ClientMetricReport</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">void</span></li>
+						<li class="tsd-signature tsd-kind-icon">metrics<wbr>Did<wbr>Receive<span class="tsd-signature-symbol">(</span>clientMetricReport<span class="tsd-signature-symbol">: </span><a href="clientmetricreport.html" class="tsd-signature-type">ClientMetricReport</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">void</span></li>
 					</ul>
 					<ul class="tsd-descriptions">
 						<li class="tsd-description">
@@ -213,7 +213,7 @@
 							<h4 class="tsd-parameters-title">Parameters</h4>
 							<ul class="tsd-parameters">
 								<li>
-									<h5>clientMetricReport: <a href="../interfaces/clientmetricreport.html" class="tsd-signature-type">ClientMetricReport</a></h5>
+									<h5>clientMetricReport: <a href="clientmetricreport.html" class="tsd-signature-type">ClientMetricReport</a></h5>
 								</li>
 							</ul>
 							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">void</span></h4>
