+++ docs/index.html
@@ -1372,6 +1372,9 @@ appVersion: string;
 						<a href="classes/cleanstoppedsessiontask.html" class="tsd-kind-icon">Clean<wbr>Stopped<wbr>Session<wbr>Task</a>
 					</li>
 					<li class=" tsd-kind-class">
+						<a href="classes/clientmetricreport.html" class="tsd-kind-icon">Client<wbr>Metric<wbr>Report</a>
+					</li>
+					<li class=" tsd-kind-class">
 						<a href="classes/clientvideostreamreceivingreport.html" class="tsd-kind-icon">Client<wbr>Video<wbr>Stream<wbr>Receiving<wbr>Report</a>
 					</li>
 					<li class=" tsd-kind-class">
@@ -1412,9 +1415,6 @@ appVersion: string;
 					</li>
 					<li class=" tsd-kind-class">
 						<a href="classes/defaultbrowserbehavior.html" class="tsd-kind-icon">Default<wbr>Browser<wbr>Behavior</a>
-					</li>
-					<li class=" tsd-kind-class">
-						<a href="classes/defaultclientmetricreport.html" class="tsd-kind-icon">Default<wbr>Client<wbr>Metric<wbr>Report</a>
 					</li>
 					<li class=" tsd-kind-class">
 						<a href="classes/defaultcontentsharecontroller.html" class="tsd-kind-icon">Default<wbr>Content<wbr>Share<wbr>Controller</a>
@@ -1466,9 +1466,6 @@ appVersion: string;
 					</li>
 					<li class=" tsd-kind-class">
 						<a href="classes/defaultsimulcastuplinkpolicy.html" class="tsd-kind-icon">Default<wbr>Simulcast<wbr>Uplink<wbr>Policy</a>
-					</li>
-					<li class=" tsd-kind-class">
-						<a href="classes/defaultstatscollector.html" class="tsd-kind-icon">Default<wbr>Stats<wbr>Collector</a>
 					</li>
 					<li class=" tsd-kind-class">
 						<a href="classes/defaulttransceivercontroller.html" class="tsd-kind-icon">Default<wbr>Transceiver<wbr>Controller</a>
@@ -1759,6 +1756,9 @@ appVersion: string;
 						<a href="classes/some.html" class="tsd-kind-icon">Some</a>
 					</li>
 					<li class=" tsd-kind-class">
+						<a href="classes/statscollector.html" class="tsd-kind-icon">Stats<wbr>Collector</a>
+					</li>
+					<li class=" tsd-kind-class">
 						<a href="classes/streammetricreport.html" class="tsd-kind-icon">Stream<wbr>Metric<wbr>Report</a>
 					</li>
 					<li class=" tsd-kind-class">
@@ -1922,9 +1922,6 @@ appVersion: string;
 					</li>
 					<li class=" tsd-kind-interface">
 						<a href="interfaces/browserbehavior.html" class="tsd-kind-icon">Browser<wbr>Behavior</a>
-					</li>
-					<li class=" tsd-kind-interface">
-						<a href="interfaces/clientmetricreport.html" class="tsd-kind-icon">Client<wbr>Metric<wbr>Report</a>
 					</li>
 					<li class=" tsd-kind-interface">
 						<a href="interfaces/connectionhealthpolicy.html" class="tsd-kind-icon">Connection<wbr>Health<wbr>Policy</a>
@@ -2099,9 +2096,6 @@ appVersion: string;
 					</li>
 					<li class=" tsd-kind-interface">
 						<a href="interfaces/simulcastuplinkpolicy.html" class="tsd-kind-icon">Simulcast<wbr>Uplink<wbr>Policy</a>
-					</li>
-					<li class=" tsd-kind-interface">
-						<a href="interfaces/statscollector.html" class="tsd-kind-icon">Stats<wbr>Collector</a>
 					</li>
 					<li class=" tsd-kind-interface">
 						<a href="interfaces/task.html" class="tsd-kind-icon">Task</a>
