+++ src/audiovideocontroller/DefaultAudioVideoController.ts
@@ -37,7 +37,7 @@ import SignalingClientVideoSubscriptionConfiguration from '../signalingclient/Si
 import SignalingClientObserver from '../signalingclientobserver/SignalingClientObserver';
 import { SdkStreamServiceType } from '../signalingprotocol/SignalingProtocol.js';
 import SimulcastLayers from '../simulcastlayers/SimulcastLayers';
-import DefaultStatsCollector from '../statscollector/DefaultStatsCollector';
+import StatsCollector from '../statscollector/StatsCollector';
 import AttachMediaInputTask from '../task/AttachMediaInputTask';
 import CleanRestartedSessionTask from '../task/CleanRestartedSessionTask';
 import CleanStoppedSessionTask from '../task/CleanStoppedSessionTask';
@@ -116,7 +116,7 @@ export default class DefaultAudioVideoController
   private preStartObserver: SignalingClientObserver | undefined;
   private mayNeedRenegotiationForSimulcastLayerChange: boolean = false;
   private maxUplinkBandwidthKbps: number;
-
+  private hasGetRTCPeerConnectionStatsDeprecationMessageBeenSent: boolean = false;
   // `connectWithPromises`, `connectWithTasks`, and `actionUpdateWithRenegotiation` all
   // contains a significant portion of asynchronous tasks, so we need to explicitly defer
   // any task operation which may be performed on the event queue that may modify
@@ -225,9 +225,14 @@ export default class DefaultAudioVideoController
    * and get the raw `RTCStatsReport` via `clientMetricReport.getRTCStatsReport()`.
    */
   getRTCPeerConnectionStats(selector?: MediaStreamTrack): Promise<RTCStatsReport> {
-    this.logger.warn(
-      'The `getRTCPeerConnectionStats()` is on its way to be deprecated. Please use the new API `clientMetricReport.getRTCStatsReport()` returned by `metricsDidReceive(clientMetricReport)` callback instead. It makes an additional call to the `getStats` API and therefore may cause slight performance degradation.'
-    );
+    /* istanbul ignore else */
+    if (!this.hasGetRTCPeerConnectionStatsDeprecationMessageBeenSent) {
+      this.logger.warn(
+        'The `getRTCPeerConnectionStats()` is on its way to be deprecated. It makes an additional call to the `getStats` API and therefore may cause slight performance degradation. Please use the new API `clientMetricReport.getRTCStatsReport()` returned by `metricsDidReceive(clientMetricReport)` callback instead.'
+      );
+      this.hasGetRTCPeerConnectionStatsDeprecationMessageBeenSent = true;
+    }
+
     if (!this.rtcPeerConnection) {
       return null;
     }
@@ -579,7 +584,7 @@ export default class DefaultAudioVideoController
     );
     this.meetingSessionContext.videosToReceive = new DefaultVideoStreamIdSet();
     this.meetingSessionContext.videosPaused = new DefaultVideoStreamIdSet();
-    this.meetingSessionContext.statsCollector = new DefaultStatsCollector(this, this.logger);
+    this.meetingSessionContext.statsCollector = new StatsCollector(this, this.logger);
     this.meetingSessionContext.connectionMonitor = new SignalingAndMetricsConnectionMonitor(
       this,
       this._realtimeController,
