+++ src/index.ts
@@ -79,7 +79,6 @@ import DefaultAudioMixController from './audiomixcontroller/DefaultAudioMixContr
 import DefaultAudioVideoController from './audiovideocontroller/DefaultAudioVideoController';
 import DefaultAudioVideoFacade from './audiovideofacade/DefaultAudioVideoFacade';
 import DefaultBrowserBehavior from './browserbehavior/DefaultBrowserBehavior';
-import DefaultClientMetricReport from './clientmetricreport/DefaultClientMetricReport';
 import DefaultContentShareController from './contentsharecontroller/DefaultContentShareController';
 import DefaultDeviceController from './devicecontroller/DefaultDeviceController';
 import DefaultDevicePixelRatioMonitor from './devicepixelratiomonitor/DefaultDevicePixelRatioMonitor';
@@ -97,7 +96,6 @@ import DefaultSessionStateController from './sessionstatecontroller/DefaultSessi
 import DefaultSigV4 from './sigv4/DefaultSigV4';
 import DefaultSignalingClient from './signalingclient/DefaultSignalingClient';
 import DefaultSimulcastUplinkPolicy from './videouplinkbandwidthpolicy/DefaultSimulcastUplinkPolicy';
-import DefaultStatsCollector from './statscollector/DefaultStatsCollector';
 import DefaultTransceiverController from './transceivercontroller/DefaultTransceiverController';
 import DefaultTranscriptionController from './transcript/DefaultTranscriptionController';
 import DefaultUserAgentParser from './useragentparser/DefaultUserAgentParser';
@@ -244,7 +242,6 @@ import SimulcastUplinkObserver from './videouplinkbandwidthpolicy/SimulcastUplin
 import SimulcastUplinkPolicy from './videouplinkbandwidthpolicy/SimulcastUplinkPolicy';
 import SimulcastVideoStreamIndex from './videostreamindex/SimulcastVideoStreamIndex';
 import SingleNodeAudioTransformDevice from './devicecontroller/SingleNodeAudioTransformDevice';
-import StatsCollector from './statscollector/StatsCollector';
 import StreamMetricReport from './clientmetricreport/StreamMetricReport';
 import SubscribeAndReceiveSubscribeAckTask from './task/SubscribeAndReceiveSubscribeAckTask';
 import TargetDisplaySize from './videodownlinkbandwidthpolicy/TargetDisplaySize';
@@ -400,7 +397,6 @@ export {
   DefaultAudioVideoController,
   DefaultAudioVideoFacade,
   DefaultBrowserBehavior,
-  DefaultClientMetricReport,
   DefaultContentShareController,
   DefaultDeviceController,
   DefaultDevicePixelRatioMonitor,
@@ -418,7 +414,6 @@ export {
   DefaultSigV4,
   DefaultSignalingClient,
   DefaultSimulcastUplinkPolicy,
-  DefaultStatsCollector,
   DefaultTransceiverController,
   DefaultTranscriptionController,
   DefaultUserAgentParser,
@@ -572,7 +567,6 @@ export {
   SimulcastVideoStreamIndex,
   SingleNodeAudioTransformDevice,
   Some,
-  StatsCollector,
   StreamMetricReport,
   SubscribeAndReceiveSubscribeAckTask,
   TargetDisplaySize,
