+++ test/task/CleanStoppedSessionTask.test.ts
@@ -17,7 +17,7 @@ import TimeoutScheduler from '../../src/scheduler/TimeoutScheduler';
 import DefaultSignalingClient from '../../src/signalingclient/DefaultSignalingClient';
 import SignalingClient from '../../src/signalingclient/SignalingClient';
 import SignalingClientConnectionRequest from '../../src/signalingclient/SignalingClientConnectionRequest';
-import DefaultStatsCollector from '../../src/statscollector/DefaultStatsCollector';
+import StatsCollector from '../../src/statscollector/StatsCollector';
 import CleanStoppedSessionTask from '../../src/task/CleanStoppedSessionTask';
 import Task from '../../src/task/Task';
 import DefaultTransceiverController from '../../src/transceivercontroller/DefaultTransceiverController';
@@ -49,7 +49,7 @@ describe('CleanStoppedSessionTask', () => {
   let signalingClient: SignalingClient;
   let request: SignalingClientConnectionRequest;
 
-  class TestStatsCollector extends DefaultStatsCollector {
+  class TestStatsCollector extends StatsCollector {
     constructor(audioVideoController: AudioVideoController, logger: Logger) {
       super(audioVideoController, logger);
     }
