+++ test/task/MonitorTask.test.ts
@@ -13,7 +13,6 @@ import ClientMetricReport from '../../src/clientmetricreport/ClientMetricReport'
 import ClientMetricReportDirection from '../../src/clientmetricreport/ClientMetricReportDirection';
 import ClientMetricReportMediaType from '../../src/clientmetricreport/ClientMetricReportMediaType';
 import ClientVideoStreamReceivingReport from '../../src/clientmetricreport/ClientVideoStreamReceivingReport';
-import DefaultClientMetricReport from '../../src/clientmetricreport/DefaultClientMetricReport';
 import StreamMetricReport from '../../src/clientmetricreport/StreamMetricReport';
 import ConnectionHealthData from '../../src/connectionhealthpolicy/ConnectionHealthData';
 import ConnectionHealthPolicyConfiguration from '../../src/connectionhealthpolicy/ConnectionHealthPolicyConfiguration';
@@ -42,7 +41,7 @@ import {
   SdkSignalFrame,
 } from '../../src/signalingprotocol/SignalingProtocol.js';
 import AudioLogEvent from '../../src/statscollector/AudioLogEvent';
-import DefaultStatsCollector from '../../src/statscollector/DefaultStatsCollector';
+import StatsCollector from '../../src/statscollector/StatsCollector';
 import VideoLogEvent from '../../src/statscollector/VideoLogEvent';
 import MonitorTask from '../../src/task/MonitorTask';
 import NoVideoDownlinkBandwidthPolicy from '../../src/videodownlinkbandwidthpolicy/NoVideoDownlinkBandwidthPolicy';
@@ -128,7 +127,7 @@ describe('MonitorTask', () => {
     }
   }
 
-  class TestStatsCollector extends DefaultStatsCollector {
+  class TestStatsCollector extends StatsCollector {
     constructor(audioVideoController: AudioVideoController, logger: Logger) {
       super(audioVideoController, logger);
     }
@@ -161,7 +160,7 @@ describe('MonitorTask', () => {
     );
     context.videoDownlinkBandwidthPolicy = new NoVideoDownlinkBandwidthPolicy();
     context.videosToReceive = context.videoDownlinkBandwidthPolicy.chooseSubscriptions().clone();
-    context.statsCollector = new DefaultStatsCollector(context.audioVideoController, logger);
+    context.statsCollector = new StatsCollector(context.audioVideoController, logger);
     context.reconnectController = new DefaultReconnectController(
       RECONNECT_TIMEOUT_MS,
       new FullJitterBackoff(
@@ -360,11 +359,11 @@ describe('MonitorTask', () => {
   });
 
   describe('metricsDidReceive', () => {
-    it('can only handle DefaultClientMetricReport with StreamMetricReport', () => {
+    it('can only handle ClientMetricReport with StreamMetricReport', () => {
       context.videoStreamIndex = new DefaultVideoStreamIndex(logger);
       task.handleSignalingClientEvent(createSignalingEventForBitrateFrame(logger));
 
-      class TestClientMetricReport implements ClientMetricReport {
+      class TestClientMetricReport extends ClientMetricReport {
         getObservableMetrics(): { [id: string]: number } {
           return;
         }
@@ -374,10 +373,11 @@ describe('MonitorTask', () => {
         }
       }
       task.metricsDidReceive(undefined);
-      const clientMetricReport = new TestClientMetricReport();
+      const clientMetricReport = new TestClientMetricReport(logger);
+      clientMetricReport.streamMetricReports = undefined;
       task.metricsDidReceive(clientMetricReport);
 
-      const defaultClientMetricReport = new DefaultClientMetricReport(logger);
+      const defaultClientMetricReport = new ClientMetricReport(logger);
       const ssrc = 6789;
       defaultClientMetricReport.streamMetricReports[ssrc] = new StreamMetricReport();
       task.metricsDidReceive(defaultClientMetricReport);
@@ -392,7 +392,7 @@ describe('MonitorTask', () => {
       streamMetricReport.mediaType = ClientMetricReportMediaType.VIDEO;
       streamMetricReport.direction = ClientMetricReportDirection.DOWNSTREAM;
 
-      const clientMetricReport = new DefaultClientMetricReport(logger);
+      const clientMetricReport = new ClientMetricReport(logger);
       clientMetricReport.streamMetricReports[1] = streamMetricReport;
       task.metricsDidReceive(clientMetricReport);
     });
@@ -420,7 +420,7 @@ describe('MonitorTask', () => {
       streamMetricReportAudio.mediaType = ClientMetricReportMediaType.VIDEO;
       streamMetricReportAudio.direction = ClientMetricReportDirection.UPSTREAM;
 
-      const clientMetricReport = new DefaultClientMetricReport(logger);
+      const clientMetricReport = new ClientMetricReport(logger);
       clientMetricReport.streamMetricReports[1] = streamMetricReport;
       clientMetricReport.streamMetricReports[56789] = streamMetricReportAudio;
       task.metricsDidReceive(clientMetricReport);
@@ -459,7 +459,7 @@ describe('MonitorTask', () => {
       streamMetricReportAudio.mediaType = ClientMetricReportMediaType.VIDEO;
       streamMetricReportAudio.direction = ClientMetricReportDirection.UPSTREAM;
 
-      const clientMetricReport = new DefaultClientMetricReport(logger);
+      const clientMetricReport = new ClientMetricReport(logger);
       clientMetricReport.streamMetricReports[1] = streamMetricReport;
       clientMetricReport.streamMetricReports[56789] = streamMetricReportAudio;
       task.metricsDidReceive(clientMetricReport);
@@ -481,7 +481,7 @@ describe('MonitorTask', () => {
       streamMetricReport.mediaType = ClientMetricReportMediaType.VIDEO;
       streamMetricReport.direction = ClientMetricReportDirection.DOWNSTREAM;
 
-      const clientMetricReport = new DefaultClientMetricReport(logger);
+      const clientMetricReport = new ClientMetricReport(logger);
       clientMetricReport.streamMetricReports[1] = streamMetricReport;
       task.metricsDidReceive(clientMetricReport);
     });
@@ -501,7 +501,7 @@ describe('MonitorTask', () => {
       streamMetricReport.mediaType = ClientMetricReportMediaType.VIDEO;
       streamMetricReport.direction = ClientMetricReportDirection.DOWNSTREAM;
 
-      const clientMetricReport = new DefaultClientMetricReport(logger);
+      const clientMetricReport = new ClientMetricReport(logger);
       clientMetricReport.streamMetricReports[1] = streamMetricReport;
       task.metricsDidReceive(clientMetricReport);
     });
@@ -509,7 +509,7 @@ describe('MonitorTask', () => {
     it('could trigger resubscription', done => {
       // eslint-disable-next-line
       type RawMetrics = any;
-      class TestClientMetricReport extends DefaultClientMetricReport {
+      class TestClientMetricReport extends ClientMetricReport {
         packetsReceived: RawMetrics = null;
         fractionLoss: RawMetrics = null;
         videoPacketSentPerSecond: RawMetrics = 1000;
@@ -597,7 +597,7 @@ describe('MonitorTask', () => {
   it('will not trigger resubscription if paused', done => {
     // eslint-disable-next-line
     type RawMetrics = any;
-    class TestClientMetricReport extends DefaultClientMetricReport {
+    class TestClientMetricReport extends ClientMetricReport {
       packetsReceived: RawMetrics = null;
       fractionLoss: RawMetrics = null;
       videoPacketSentPerSecond: RawMetrics = 1000;
@@ -688,7 +688,7 @@ describe('MonitorTask', () => {
   it('will trigger resubscription if paused and resumed', done => {
     // eslint-disable-next-line
     type RawMetrics = any;
-    class TestClientMetricReport extends DefaultClientMetricReport {
+    class TestClientMetricReport extends ClientMetricReport {
       packetsReceived: RawMetrics = null;
       fractionLoss: RawMetrics = null;
       videoPacketSentPerSecond: RawMetrics = 1000;
